requirements.txt
streamlit
pandas
plotly
python-docx
google-generativeai
openai
dotenv